Refs #6: Refactor all the domain and application packages | Add build command in Makefile
