+++ servers/game/src/main/kotlin/fr/rob/game/Main.kt
@@ -3,10 +3,10 @@ package fr.rob.game
 import fr.rob.core.AbstractModule
 import fr.rob.core.BaseApplication
 import fr.rob.core.initiator.Initiator
-import fr.rob.game.application.network.GameServerManager
-import fr.rob.game.application.setup.AppSetup
-import fr.rob.game.application.setup.tasks.TaskAuthCollectJWTPublicKey
-import fr.rob.game.application.setup.tasks.TaskLoadServerConfig
+import fr.rob.game.domain.network.GameServerManager
+import fr.rob.game.domain.setup.AppSetup
+import fr.rob.game.domain.setup.tasks.TaskAuthCollectJWTPublicKey
+import fr.rob.game.domain.setup.tasks.TaskLoadServerConfig
 import fr.rob.game.domain.network.ServerManager
 import fr.rob.game.domain.setup.Setup
 import fr.rob.game.infrastructure.config.ResourceManager
