+++ servers/game/src/main/kotlin/fr/rob/game/domain/authentication/AuthenticationProcess.kt
@@ -1,7 +1,7 @@
 package fr.rob.game.domain.authentication
 
 import fr.rob.core.auth.jwt.JWTDecoderInterface
-import fr.rob.game.domain.network.Session
+import fr.rob.game.domain.network.session.Session
 import java.lang.Exception
 
 /**
