+++ servers/game/src/main/kotlin/fr/rob/game/domain/network/GameServer.kt
@@ -3,6 +3,7 @@ package fr.rob.game.domain.network
 import fr.rob.game.domain.log.LoggerInterface
 import fr.rob.game.domain.opcode.ClientOpcodeHandler
 import fr.rob.game.domain.network.exception.SessionNotFoundException
+import fr.rob.game.domain.network.session.Session
 
 open class GameServer(logger: LoggerInterface) {
