+++ servers/game/src/main/kotlin/fr/rob/game/domain/network/netty/NettyGameServer.kt
@@ -1,4 +1,4 @@
-package fr.rob.game.application.network
+package fr.rob.game.domain.network.netty
 
 import fr.rob.game.SSL_ENABLED
 import fr.rob.game.domain.log.LoggerInterface
@@ -11,7 +11,7 @@ import io.netty.channel.nio.NioEventLoopGroup
 import io.netty.channel.socket.nio.NioServerSocketChannel
 import kotlin.concurrent.thread
 
-class GameServer(private val port: Int, logger: LoggerInterface) : GameServer(logger) {
+class NettyGameServer(private val port: Int, logger: LoggerInterface) : GameServer(logger) {
 
     private val bootstrap: ServerBootstrap = ServerBootstrap()
 
@@ -24,7 +24,7 @@ class GameServer(private val port: Int, logger: LoggerInterface) : GameServer(lo
             .childOption(ChannelOption.SO_KEEPALIVE, true)
 
         // Our handler
-        bootstrap.childHandler(ServerInitializer(this, SSL_ENABLED))
+        bootstrap.childHandler(NettyServerInitializer(this, SSL_ENABLED))
 
         val channel: ChannelFuture = bootstrap.bind(port).sync()
