+++ servers/game/src/main/kotlin/fr/rob/game/domain/network/session/AppSession.kt
@@ -1,13 +1,12 @@
-package fr.rob.game.application.network
+package fr.rob.game.domain.network.session
 
-import fr.rob.game.domain.network.GameServer
-import fr.rob.game.domain.network.Session
+import fr.rob.game.domain.network.netty.NettyGameServer
 import fr.rob.game.domain.network.packet.Packet
 import io.netty.channel.Channel
 import io.netty.util.NetUtil
 import java.net.InetSocketAddress
 
-class AppSession(gameServer: GameServer, private val channel: Channel) : Session(gameServer) {
+class AppSession(nettyGameServer: NettyGameServer, private val channel: Channel) : Session(nettyGameServer) {
 
     override fun getIp(): String {
         val address = channel.remoteAddress() as InetSocketAddress
