+++ servers/game/src/main/kotlin/fr/rob/game/domain/opcode/OpcodeFunction.kt
@@ -1,6 +1,6 @@
 package fr.rob.game.domain.opcode
 
-import fr.rob.game.domain.network.Session
+import fr.rob.game.domain.network.session.Session
 import fr.rob.game.domain.network.packet.Packet
 
 abstract class OpcodeFunction {
