+++ servers/game/src/main/kotlin/fr/rob/game/domain/opcode/OpcodeHandler.kt
@@ -1,7 +1,7 @@
 package fr.rob.game.domain.opcode
 
 import fr.rob.game.domain.log.LoggerInterface
-import fr.rob.game.domain.network.Session
+import fr.rob.game.domain.network.session.Session
 import fr.rob.game.domain.network.packet.Packet
 
 open class OpcodeHandler(private val logger: LoggerInterface) {
