+++ servers/game/src/main/kotlin/fr/rob/game/infrastructure/database/event/DatabaseEvent.kt
@@ -1,6 +1,6 @@
 package fr.rob.game.infrastructure.database.event
 
-import fr.rob.game.infrastructure.event.Event
+import fr.rob.game.domain.event.Event
 
 abstract class DatabaseEvent(val sql: String, val time: Long) : Event() {
 }
