+++ servers/game/src/test/kotlin/fr/rob/test/domain/opcode/OpcodeHandlerTest.kt
@@ -1,7 +1,7 @@
 package fr.rob.test.domain.opcode
 
-import fr.rob.game.domain.opcode.OpcodeHandler
 import fr.rob.game.domain.network.GameServer
+import fr.rob.game.domain.opcode.OpcodeHandler
 import fr.rob.game.domain.network.packet.Packet
 import fr.rob.game.sandbox.SandboxProtos
 import fr.rob.test.BaseTest
