+++ servers/game/src/test/kotlin/fr/rob/test/opcode/HandlingOpcodeWithProtoAsMessageOpcode.kt
@@ -1,8 +1,8 @@
 package fr.rob.test.opcode
 
 import com.google.protobuf.Message
-import fr.rob.game.application.opcode.ProtobufOpcodeFunction
-import fr.rob.game.domain.network.Session
+import fr.rob.game.domain.network.session.Session
+import fr.rob.game.domain.opcode.ProtobufOpcodeFunction
 import fr.rob.game.sandbox.SandboxProtos
 
 class HandlingOpcodeWithProtoAsMessageOpcode : ProtobufOpcodeFunction() {
