+++ servers/game/src/test/kotlin/fr/rob/test/sandbox/network/NISession.kt
@@ -1,7 +1,7 @@
 package fr.rob.test.sandbox.network
 
 import fr.rob.game.domain.network.GameServer
-import fr.rob.game.domain.network.Session
+import fr.rob.game.domain.network.session.Session
 import fr.rob.game.domain.network.packet.Packet
 
 class NISession(gameServer: GameServer) : Session(gameServer) {
