+++ servers/game/src/test/kotlin/fr/rob/test/sandbox/opcode/DoNothingOpcodeFunction.kt
@@ -1,6 +1,6 @@
 package fr.rob.test.sandbox.opcode
 
-import fr.rob.game.domain.network.Session
+import fr.rob.game.domain.network.session.Session
 
 class DoNothingOpcodeFunction(subject: Any) : OpcodeHandlerTestSubject(subject) {
